src/electrai/requirements.txt
numpy<2
scikit-learn==1.7.2
torch==1.11.0+cu113 
torchvision==0.12.0+cu113
pyscf==2.10.0
pyyaml=6.0.3
monty==2025.3.3
pymatgen==2025.10.7

